requirements-eval.txt
# Core agent dependencies (already in requirements.txt if it exists)
anthropic>=0.40.0
rank-bm25>=0.2.2

# Eval framework
pydantic>=2.0.0
pyyaml>=6.0
